+++ src/NS3D/TerrainShader.cpp
@@ -27,7 +27,7 @@ namespace NS3D {
         location_blendMap          = getUniformLocation("blendMap"         );
         location_clipPlane         = getUniformLocation("clipPlane"        );
 
-        for (int i = 0; i < MAX_LIGHTS; i++)
+        for (size_t i = 0; i < MAX_LIGHTS; i++)
         {
             location_lightPosition[i] = getUniformLocation("lightPosition["+std::to_string(i)+"]");
             location_lightColor[i]    = getUniformLocation("lightColor["+std::to_string(i)+"]"   );
@@ -58,7 +58,7 @@ namespace NS3D {
     }
     void TerrainShader::SetLights(const std::vector<const Light*>& lights)
     {
-        for (int i = 0; i < MAX_LIGHTS; i++)
+        for (size_t i = 0; i < MAX_LIGHTS; i++)
         {
             if (i < lights.size())
             {
